requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pinecone-client==2.2.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2
pypdf2==3.0.1
python-docx==1.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
aiofiles==23.2.1
httpx==0.25.2 